agents-backend/agents/requirements.txt
requests
psycopg2-binary
pandas
pyyaml
defog
snowflake-connector-python
slack_sdk
pandas
resend
sqlalchemy
scipy
tiktoken
fastapi
scikit-survival
numpy
google-cloud-storage
databricks-sql-connector
pika
hypercorn
colorama
matplotlib
seaborn